# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OF.ru",
        "version": "1.0.0"
    },
    "paths": {
...
        "/v1/overlap-types/{id}": {
            "get": {
                "tags": [
                    "OverlapType"
                ],
                "description": "Display the specified Item.",
                "operationId": "0e5495d46c6a7fa7dc09805cbf88c6be",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/v1/rent-block-taxes/{id}": {
            "get": {
                "tags": [
                    "RentBlockTax"
                ],
                "description": "Display the specified Item.",
                "operationId": "364fb8ef1a8c60754804a10ceba19e22",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/v1/target-sales-types/{id}": {
            "get": {
                "tags": [
                    "TargetSalesType"
                ],
                "description": "Display the specified Item.",
                "operationId": "b1f7187dac4d4cf0969f1c0c54b2916c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
        "/v1/ventilations/{id}": {
            "get": {
                "tags": [
                    "Ventilation"
                ],
                "description": "Display the specified Item.",
                "operationId": "b918125cae887503b649325e63f92602",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
...
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/",
            "description": "Server parameters"
        }
    ]
}